requirements.txt
# Install required packages for DHT11 sensor
adafruit-circuitpython-dht
board
RPi.GPIO
Adafruit_DHT
